Changed Syndication Tests to Test Sync APIs.
